// frontend/web/src/services/API/proserve-wb-packaging-api/models/GetPipelinesAllowedBuildTypesResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface GetPipelinesAllowedBuildTypesResponse
 */
export interface GetPipelinesAllowedBuildTypesResponse {
    /**
     * 
     * @type {Array<string>}
     * @memberof GetPipelinesAllowedBuildTypesResponse
     */
    pipelinesAllowedBuildTypes: Array<string>;
}

/**
 * Check if a given object implements the GetPipelinesAllowedBuildTypesResponse interface.
 */
export function instanceOfGetPipelinesAllowedBuildTypesResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "pipelinesAllowedBuildTypes" in value;

    return isInstance;
}

export function GetPipelinesAllowedBuildTypesResponseFromJSON(json: any): GetPipelinesAllowedBuildTypesResponse {
    return GetPipelinesAllowedBuildTypesResponseFromJSONTyped(json, false);
}

export function GetPipelinesAllowedBuildTypesResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetPipelinesAllowedBuildTypesResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'pipelinesAllowedBuildTypes': json['pipelines_allowed_build_types'],
    };
}

export function GetPipelinesAllowedBuildTypesResponseToJSON(value?: GetPipelinesAllowedBuildTypesResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'pipelines_allowed_build_types': value.pipelinesAllowedBuildTypes,
    };
}