// frontend/web/src/services/API/proserve-wb-packaging-api/models/GetRecipesVersionsResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { RecipeVersionSummary } from './RecipeVersionSummary';
import {
    RecipeVersionSummaryFromJSON,
    RecipeVersionSummaryFromJSONTyped,
    RecipeVersionSummaryToJSON,
} from './RecipeVersionSummary';

/**
 * 
 * @export
 * @interface GetRecipesVersionsResponse
 */
export interface GetRecipesVersionsResponse {
    /**
     * 
     * @type {Array<RecipeVersionSummary>}
     * @memberof GetRecipesVersionsResponse
     */
    recipesVersionsSummary: Array<RecipeVersionSummary>;
}

/**
 * Check if a given object implements the GetRecipesVersionsResponse interface.
 */
export function instanceOfGetRecipesVersionsResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "recipesVersionsSummary" in value;

    return isInstance;
}

export function GetRecipesVersionsResponseFromJSON(json: any): GetRecipesVersionsResponse {
    return GetRecipesVersionsResponseFromJSONTyped(json, false);
}

export function GetRecipesVersionsResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetRecipesVersionsResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'recipesVersionsSummary': ((json['recipes_versions_summary'] as Array<any>).map(RecipeVersionSummaryFromJSON)),
    };
}

export function GetRecipesVersionsResponseToJSON(value?: GetRecipesVersionsResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'recipes_versions_summary': ((value.recipesVersionsSummary as Array<any>).map(RecipeVersionSummaryToJSON)),
    };
}