// frontend/web/src/services/API/proserve-wb-packaging-api/models/RecipeVersionSummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface RecipeVersionSummary
 */
export interface RecipeVersionSummary {
    /**
     * Recipe ID
     * @type {string}
     * @memberof RecipeVersionSummary
     */
    recipeId: string;
    /**
     * Recipe version ID
     * @type {string}
     * @memberof RecipeVersionSummary
     */
    recipeVersionId: string;
    /**
     * Recipe version Name
     * @type {string}
     * @memberof RecipeVersionSummary
     */
    recipeVersionName: string;
    /**
     * Recipe Name
     * @type {string}
     * @memberof RecipeVersionSummary
     */
    recipeName: string;
}

/**
 * Check if a given object implements the RecipeVersionSummary interface.
 */
export function instanceOfRecipeVersionSummary(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "recipeId" in value;
    isInstance = isInstance && "recipeVersionId" in value;
    isInstance = isInstance && "recipeVersionName" in value;
    isInstance = isInstance && "recipeName" in value;

    return isInstance;
}

export function RecipeVersionSummaryFromJSON(json: any): RecipeVersionSummary {
    return RecipeVersionSummaryFromJSONTyped(json, false);
}

export function RecipeVersionSummaryFromJSONTyped(json: any, ignoreDiscriminator: boolean): RecipeVersionSummary {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'recipeId': json['recipeId'],
        'recipeVersionId': json['recipeVersionId'],
        'recipeVersionName': json['recipeVersionName'],
        'recipeName': json['recipeName'],
    };
}

export function RecipeVersionSummaryToJSON(value?: RecipeVersionSummary | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'recipeId': value.recipeId,
        'recipeVersionId': value.recipeVersionId,
        'recipeVersionName': value.recipeVersionName,
        'recipeName': value.recipeName,
    };
}